move app controller to separate file
