+++ frontend/config.toml
@@ -3,7 +3,7 @@ docked = false
 
 [[fetch]]
 from = 'controllers'
-files = ['__init__.py', 'click_rotation_controller.py', 'drag_drop_grid_controller.py', 'drag_rotation_controller.py', 'image_grid_controller.py', 'rotating_images_controller.py', 'rotation_controller.py']
+files = ['__init__.py', 'app.py', 'click_rotation_controller.py', 'drag_drop_grid_controller.py', 'drag_rotation_controller.py', 'image_grid_controller.py', 'rotating_images_controller.py', 'rotation_controller.py']
 to_folder = 'controllers'
