+++ frontend/controllers/__init__.py
@@ -1,4 +1,2 @@
-"""Provides the controllers for the frontend app."""
-from .drag_drop_grid_controller import DragDropGridController
-from .image_grid_controller import ImageGridController
-from .rotating_images_controller import RotatingImagesController
+"""Provides the frontend app controller."""
+from .app import App
